backend/pyproject.toml
[project]
name = "reduse"
version = "0.1.0"
description = ""
authors = [
    {name = "Mycsina", email = "andremacardoso@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv",
    "beautifulsoup4",
    "fastapi[standard]",
    "playwright",
    "google-generativeai",
    "beanie",
    "motor>=3.6.0",
    "pymongo",
    "groq",
    "pydantic-settings",
    "tqdm",
    "httpx",
    "apscheduler",
    "certifi",
    "sse-starlette",
    "scikit-learn",
    "crawlee",
    "umap-learn>=0.5.7",
    "tenacity>=9.1.2",
    "fastapi-users[beanie,oauth]>=14.0.1",
]

[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "isort>=6.0.0",
    "fastapi-mcp>=0.1.3",
]

[tool.uv]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "vulture>=2.14",
]

[tool.vulture]
exclude = ["venv/"]
min_confidence = 65
sort_by_size = true
